requirements.txt
attrs==17.3.0
Automat==0.6.0
constantly==15.1.0
hyperlink==17.3.1
incremental==17.5.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
six==1.11.0
Twisted==17.9.0
zope.interface==4.4.3
